backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
SQLAlchemy==2.0.21
PyPDF2==3.0.1
python-docx==0.8.11
python-magic==0.4.27
spacy==3.6.1
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
APScheduler==3.10.4
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.20.1
Werkzeug==2.3.7